Remove redundant .astype(int) casts; make worm divergence integer
Now that d preserves integer dtype, the .astype(int) casts compensating for its
old float promotion are redundant no-ops: gauge_transform's d(k) (k is guarded
integer) and ExactUpdate's d(change_z) (change_z is an int 0-form).  Drop them.

The worldline ClassicWorm's `divergence` was float, so the njit kernel did
`m_2d[...] += delta_m` with m now int and delta_m a ±1.0 float (a lossless but
ugly int+=float).  Make divergence int so the per-link update is a clean
int+=int; change_S stays float via change_link (m - δv/W) regardless.

+++ supervillain/generator/villain/exact.py
@@ -95,7 +95,7 @@ class ExactUpdate(ReadWriteable, Generator):
             change_z[0, *color] = self.rng.choice(self.zs, len(color[0]))
 
             # To keep dn=0 we let the change in n be given by d(z), so that d(change_n) = 0, since it is d^2(z).
-            change_n = d(change_z).astype(int)
+            change_n = d(change_z)
 
             dS_link = (
                 -2 * np.pi * S.kappa * change_n
@@ -117,7 +117,7 @@ class ExactUpdate(ReadWriteable, Generator):
 
             # Finally, we update the n where the change is accepted.
             change_z[0, *color] *= accepted
-            n = n + d(change_z).astype(int)
+            n = n + d(change_z)
 
         sites = self.Lattice.cells_of_degree[0]
         self.proposed += sites
